Fix some obsolete references

+++ lib/model/user.model.dart
@@ -1,5 +1,5 @@
 class User {
-  final int id;
+  final BigInt id;
   final String name;
   final String email;
   final String token;
